マイグレーションファイル再修正

+++ database/migrations/2017_09_29_092337_create_teiki_account_ledgers.php
@@ -133,7 +133,7 @@ class CreateTeikiAccountLedgers extends Migration
             $table->char("spare_1", 190);
             $table->integer("monthly_id")->index();
             $table->integer("subject_code")->index();
-            $table->integer("key_account_number")->index();
+            $table->double("key_account_number")->index();
             $table->timestamps("");
         });
     }
